Updated Basic Command Template for elevator command and creating command objects in robot container.
